chatbot/requirements.txt
# ============================================
# FAQ Chatbot Dependencies
# ============================================
# Install all dependencies with: pip install -r requirements.txt

# === LlamaIndex Core ===
# The main framework for building LLM applications with your data
llama-index==0.10.68
llama-index-core==0.10.68

# === LlamaIndex Integrations ===
# HuggingFace embeddings - converts text to vectors (runs locally)
llama-index-embeddings-huggingface==0.2.3

# Anthropic integration - connects to Claude AI
llama-index-llms-anthropic==0.1.16

# File readers - handles PDF, TXT, and other file types
llama-index-readers-file==0.1.33

# ChromaDB integration - connects LlamaIndex to ChromaDB
llama-index-vector-stores-chroma==0.1.10

# === Vector Database ===
# ChromaDB - stores document embeddings locally
chromadb==0.4.24

# === Document Processing ===
# PyPDF - reads PDF files
pypdf==4.2.0

# === Utilities ===
# Load environment variables from .env file
python-dotenv==1.0.1

# Anthropic SDK - official Claude API client
anthropic==0.29.0
